move store locator contexts
