+++ packages/template-retail-react-app/app/components/store-locator/index.tsx
@@ -7,4 +7,3 @@
 
 export {StoreLocator} from './main'
 export {StoreLocatorModal} from './modal'
-export {StoreLocatorProvider} from './provider'
