+++ packages/template-retail-react-app/app/contexts/store-locator-provider.test.jsx
@@ -9,7 +9,7 @@ import {render, act} from '@testing-library/react'
 import {
     StoreLocatorProvider,
     StoreLocatorContext
-} from '@salesforce/retail-react-app/app/components/store-locator/provider'
+} from '@salesforce/retail-react-app/app/contexts/store-locator-provider'
 
 describe('StoreLocatorProvider', () => {
     const mockConfig = {
