+++ packages/template-retail-react-app/app/utils/test-utils.js
@@ -20,11 +20,10 @@ import {withReactQuery} from '@salesforce/pwa-kit-react-sdk/ssr/universal/compon
 import fallbackMessages from '@salesforce/retail-react-app/app/static/translations/compiled/en-GB.json'
 import mockConfig from '@salesforce/retail-react-app/config/mocks/default'
 // Contexts
-import {CurrencyProvider, MultiSiteProvider} from '@salesforce/retail-react-app/app/contexts'
+import {CurrencyProvider, MultiSiteProvider, StoreLocatorProvider} from '@salesforce/retail-react-app/app/contexts'
 
 import {createUrlTemplate} from '@salesforce/retail-react-app/app/utils/url'
 import {getSiteByReference} from '@salesforce/retail-react-app/app/utils/site-utils'
-import {StoreLocatorProvider} from '@salesforce/retail-react-app/app/components/store-locator'
 import {
     STORE_LOCATOR_RADIUS,
     STORE_LOCATOR_RADIUS_UNIT,
